+++ common/perf-test-core/src/main/java/com/azure/perf/test/core/PerfStressTest.java
@@ -10,24 +10,20 @@ import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslContextBuilder;
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
 import reactor.core.publisher.Mono;
-import reactor.netty.http.client.HttpClientResponse;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
-import java.util.ArrayList;
 import javax.net.ssl.SSLException;
 
 /**
  * Represents the abstraction of a Performance test class.
  *
  * <p>
- *     The performance test class needs to extend this class. The test class should override {@link PerfStressTest#run()}
- *     and {@link PerfStressTest#runAsync()} methods and the synchronous and asynchronous test logic respectively.
- *     To add any test setup and logic the test class should override {@link PerfStressTest#globalSetupAsync()}
- *     and {@link PerfStressTest#globalCleanupAsync()} methods .
+ * The performance test class needs to extend this class. The test class should override {@link PerfStressTest#run()}
+ * and {@link PerfStressTest#runAsync()} methods and the synchronous and asynchronous test logic respectively.
+ * To add any test setup and logic the test class should override {@link PerfStressTest#globalSetupAsync()}
+ * and {@link PerfStressTest#globalCleanupAsync()} methods .
  * </p>
- *
- *
  * @param <TOptions> the options configured for the test.
  */
 public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
@@ -45,6 +41,7 @@ public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
     /**
      * Creates an instance of performance test.
      * @param options the options configured for the test.
+     * @throws IllegalStateException if SSL context cannot be created.
      */
     public PerfStressTest(TOptions options) {
         this.options = options;
@@ -54,19 +51,17 @@ public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
         if (options.isInsecure()) {
             try {
                 sslContext = SslContextBuilder.forClient()
-                    .trustManager(InsecureTrustManagerFactory.INSTANCE)
-                    .build();
-            }
-            catch (SSLException e) {
+                        .trustManager(InsecureTrustManagerFactory.INSTANCE)
+                        .build();
+            } catch (SSLException e) {
                 throw new IllegalStateException(e);
             }
 
             reactor.netty.http.client.HttpClient nettyHttpClient = reactor.netty.http.client.HttpClient.create()
-                .secure(sslContextSpec -> sslContextSpec.sslContext(sslContext));
-            
+                    .secure(sslContextSpec -> sslContextSpec.sslContext(sslContext));
+
             httpClient = new NettyAsyncHttpClientBuilder(nettyHttpClient).build();
-        }
-        else {
+        } else {
             sslContext = null;
             httpClient = null;
         }
@@ -74,25 +69,27 @@ public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
         if (options.getTestProxy() != null) {
             if (options.isInsecure()) {
                 recordPlaybackHttpClient = reactor.netty.http.client.HttpClient.create()
-                    .secure(sslContextSpec -> sslContextSpec.sslContext(sslContext));
-            }
-            else {
+                        .secure(sslContextSpec -> sslContextSpec.sslContext(sslContext));
+            } else {
                 recordPlaybackHttpClient = reactor.netty.http.client.HttpClient.create();
             }
 
             testProxyPolicy = new TestProxyPolicy(options.getTestProxy());
             policies = Arrays.asList(testProxyPolicy);
-        }
-        else {
+        } else {
             recordPlaybackHttpClient = null;
             testProxyPolicy = null;
             policies = null;
         }
     }
 
-    // Attempts to configure a ClientBuilder using reflection.  If a ClientBuilder does not follow the standard convention,
-    // it can be configured manually using the "httpClient" and "policies" fields.
-    protected void ConfigureClientBuilder(Object clientBuilder) {
+    /**
+     * Attempts to configure a ClientBuilder using reflection.  If a ClientBuilder does not follow the standard convention,
+     * it can be configured manually using the "httpClient" and "policies" fields.
+     * @param clientBuilder The client builder.
+     * @throws IllegalStateException If reflective access to get httpClient or addPolicy methods fail.
+     */
+    protected void configureClientBuilder(Object clientBuilder) {
         if (httpClient != null || policies != null) {
             Class<?> clientBuilderClass = clientBuilder.getClass();
 
@@ -101,17 +98,16 @@ public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
                     Method httpClientMethod = clientBuilderClass.getMethod("httpClient", HttpClient.class);
                     httpClientMethod.invoke(clientBuilder, httpClient);
                 }
-    
+
                 if (policies != null) {
                     Method addPolicyMethod = clientBuilderClass.getMethod("addPolicy", HttpPipelinePolicy.class);
                     for (HttpPipelinePolicy policy : policies) {
                         addPolicyMethod.invoke(clientBuilder, policy);
                     }
                 }
-            }
-            catch (ReflectiveOperationException e) {
+            } catch (ReflectiveOperationException e) {
                 throw new IllegalStateException(e);
-            } 
+            }
         }
     }
 
@@ -131,27 +127,30 @@ public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
         return Mono.empty();
     }
 
+    /**
+     * Records responses and starts async tests in playback mode.
+     * @return An empty {@link Mono}.
+     */
     public Mono<Void> recordAndStartPlaybackAsync() {
         return startRecordingAsync()
-            .doOnSuccess(x -> {
-                testProxyPolicy.setRecordingId(recordingId);
-                testProxyPolicy.setMode("record");
-            })
-            // Must use Mono.defer() to ensure fields are set from prior requests
-            .then(Mono.defer(() -> runSyncOrAsync()))
-            .then(Mono.defer(() -> stopRecordingAsync()))
-            .then(Mono.defer(() -> startPlaybackAsync()))
-            .doOnSuccess(x -> {
-                testProxyPolicy.setRecordingId(recordingId);
-                testProxyPolicy.setMode("playback");
-            });
+                .doOnSuccess(x -> {
+                    testProxyPolicy.setRecordingId(recordingId);
+                    testProxyPolicy.setMode("record");
+                })
+                // Must use Mono.defer() to ensure fields are set from prior requests
+                .then(Mono.defer(() -> runSyncOrAsync()))
+                .then(Mono.defer(() -> stopRecordingAsync()))
+                .then(Mono.defer(() -> startPlaybackAsync()))
+                .doOnSuccess(x -> {
+                    testProxyPolicy.setRecordingId(recordingId);
+                    testProxyPolicy.setMode("playback");
+                });
     }
 
     private Mono<Void> runSyncOrAsync() {
         if (options.isSync()) {
             return Mono.empty().then().doOnSuccess(x -> run());
-        }
-        else {
+        } else {
             return runAsync();
         }
     }
@@ -167,20 +166,24 @@ public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
      */
     public abstract Mono<Void> runAsync();
 
+    /**
+     * Stops playback tests.
+     * @return An empty {@link Mono}.
+     */
     public Mono<Void> stopPlaybackAsync() {
         return recordPlaybackHttpClient
-            .headers(h -> {
-                h.set("x-recording-id", recordingId);
-                h.set("x-purge-inmemory-recording", Boolean.toString(true));
-            })
-            .post()
-            .uri(options.getTestProxy().resolve("/playback/stop"))
-            .response()
-            .doOnSuccess(response -> {
-                testProxyPolicy.setMode(null);
-                testProxyPolicy.setRecordingId(null);
-            })
-            .then();
+                .headers(h -> {
+                    h.set("x-recording-id", recordingId);
+                    h.set("x-purge-inmemory-recording", Boolean.toString(true));
+                })
+                .post()
+                .uri(options.getTestProxy().resolve("/playback/stop"))
+                .response()
+                .doOnSuccess(response -> {
+                    testProxyPolicy.setMode(null);
+                    testProxyPolicy.setRecordingId(null);
+                })
+                .then();
     }
 
     /**
@@ -201,33 +204,33 @@ public abstract class PerfStressTest<TOptions extends PerfStressOptions> {
 
     private Mono<Void> startRecordingAsync() {
         return recordPlaybackHttpClient
-            .post()
-            .uri(options.getTestProxy().resolve("/record/start"))
-            .response()
-            .doOnNext(response -> {
-                recordingId = response.responseHeaders().get("x-recording-id");
-            })
-            .then();
+                .post()
+                .uri(options.getTestProxy().resolve("/record/start"))
+                .response()
+                .doOnNext(response -> {
+                    recordingId = response.responseHeaders().get("x-recording-id");
+                })
+                .then();
     }
 
     private Mono<Void> stopRecordingAsync() {
         return recordPlaybackHttpClient
-            .headers(h -> h.set("x-recording-id", recordingId))
-            .post()
-            .uri(options.getTestProxy().resolve("/record/stop"))
-            .response()
-            .then();
+                .headers(h -> h.set("x-recording-id", recordingId))
+                .post()
+                .uri(options.getTestProxy().resolve("/record/stop"))
+                .response()
+                .then();
     }
 
     private Mono<Void> startPlaybackAsync() {
         return recordPlaybackHttpClient
-            .headers(h -> h.set("x-recording-id", recordingId))
-            .post()
-            .uri(options.getTestProxy().resolve("/playback/start"))
-            .response()
-            .doOnNext(response -> {
-                recordingId = response.responseHeaders().get("x-recording-id");
-            })
-            .then();
+                .headers(h -> h.set("x-recording-id", recordingId))
+                .post()
+                .uri(options.getTestProxy().resolve("/playback/start"))
+                .response()
+                .doOnNext(response -> {
+                    recordingId = response.responseHeaders().get("x-recording-id");
+                })
+                .then();
     }
 }