+++ sdk/core/azure-core-http-netty/CHANGELOG.md
@@ -3,6 +3,24 @@
 ## 1.11.0-beta.1 (Unreleased)
 
 
+## 1.10.2 (2021-08-06)
+
+### Features Added
+
+- Added support for setting per-call response timeouts by passing `azure-response-timeout` in the request's `Context`. ([#23244](https://github.com/Azure/azure-sdk-for-java/pull/23244))
+
+### Fixed
+
+- Fixed a bug where `NullPointerException` would be thrown when `HttpClientOptions` was used in `NettyAsyncClientProvider`
+  without a connection pool size configured. ([#23357](https://github.com/Azure/azure-sdk-for-java/pull/23357))
+- Fixed a bug where the fix `CONNECT` request made after the creation of an `HttpClient` instance would time out. ([#22661](https://github.com/Azure/azure-sdk-for-java/pull/22661))
+
+### Dependency Updates
+
+- Upgraded `azure-core` from `1.18.0` to `1.19.0`.
+- Upgraded Netty from `4.1.65.Final` to `4.1.66.Final`.
+- Upgraded Reactor Netty from `1.0.8` to `1.0.9`.
+- 
 ## 1.10.1 (2021-07-01)
 
 ### Features Added