+++ sdk/core/azure-core-serializer-json-gson/CHANGELOG.md
@@ -3,6 +3,13 @@
 ## 1.2.0-beta.1 (Unreleased)
 
 
+## 1.1.5 (2021-08-06)
+
+### Dependency Updates
+
+- Upgraded `azure-core` from `1.18.0` to `1.19.0`.
+- Upgraded `gson` from `2.8.6` to `2.8.7`.
+
 ## 1.1.4 (2021-07-01)
 
 ### Dependency Updates